requirements.txt
anthropic
fastapi
httpx
playwright
pydantic
PyJWT
python-dotenv
pyyaml
uvicorn[standard]
